Small fixes, documentation
